Routes has been refactored. Their components has been adjusted to eslint standard. But, routes do not working now.
