+++ Unity/Exoplanets/react/README.md
@@ -4,7 +4,7 @@ This project was generated by `@reactunity/create`.
 
 Here are some useful scripts to get you started:
 
-- `yarn` - Install dependencies
-- `yarn build` - Build the project at the directory specified at `.env` file
-- `yarn start` - Start the local dev server at the port written in `.env` file
+- `pnpm install` - Install dependencies
+- `pnpm build` - Build the project at the directory specified at `.env` file
+- `pnpm start` - Start the local dev server at the port written in `.env` file
