+++ Unity/Exoplanets/react/src/pages/Exoplanets.tsx
@@ -1,15 +1,14 @@
 import { useNavigate } from 'react-router-dom';
 
-const MiCuenta = () => {
+export default function Exoplanets() {
   const navigate = useNavigate();
 
   return (
     <div>
-      <h1>Mi Cuenta</h1>
+      <h1>Lista de Exoplanetas</h1>
       <button onClick={() => navigate(-1)}>Retroceder</button>
       <button onClick={() => navigate('/')}>Volver al Inicio</button>
+
     </div>
   );
-};
-
-export default MiCuenta;
+}
