+++ Unity/Exoplanets/react/src/pages/Maximized.tsx
@@ -1,16 +1,13 @@
 import { useNavigate } from 'react-router-dom';
 
-const Exoplanetas = () => {
+export default function Maximized() {
   const navigate = useNavigate();
 
   return (
     <div>
-      <h1>Lista de Exoplanetas</h1>
+      <h1>Vista 3D Maximizada</h1>
       <button onClick={() => navigate(-1)}>Retroceder</button>
       <button onClick={() => navigate('/')}>Volver al Inicio</button>
-
     </div>
   );
-};
-
-export default Exoplanetas;
+}
