+++ Unity/Exoplanets/react/src/pages/MyProfile.tsx
@@ -1,15 +1,13 @@
 import { useNavigate } from 'react-router-dom';
 
-const Maximizado = () => {
+export default function MyProfile() {
   const navigate = useNavigate();
 
   return (
     <div>
-      <h1>Vista 3D Maximizada</h1>
+      <h1>Mi Cuenta</h1>
       <button onClick={() => navigate(-1)}>Retroceder</button>
       <button onClick={() => navigate('/')}>Volver al Inicio</button>
     </div>
   );
-};
-
-export default Maximizado;
+}
