+++ Unity/Exoplanets/react/src/pages/Stars.tsx
@@ -1,17 +1,12 @@
 import { useNavigate } from 'react-router-dom';
 
-const Inicio = () => {
+export default function Stars() {
   const navigate = useNavigate();
-
   return (
     <div>
-      <h1>Página de Inicio</h1>
-      <button onClick={() => navigate('/exoplanetas')}>Ir a Exoplanetas</button>
-      <button onClick={() => navigate('/estrellas')}>Ir a Estrellas</button>
+      <h1>Crear Constelación</h1>
       <button onClick={() => navigate(-1)}>Retroceder</button>
       <button onClick={() => navigate('/')}>Volver al Inicio</button>
     </div>
   );
-};
-
-export default Inicio;
+}
